backend/requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
pymupdf
pytesseract
pillow
spacy
torch
transformers
faiss-cpu